CCM-5752: wip, move test files to match project structure
