+++ src/__tests__/app/create-template/main-server-action.test.ts
@@ -2,9 +2,9 @@ import {
   handleForm as nhsAppHandleForm,
   handleFormBack as nhsAppHandleFormBack,
 } from '@forms/ReviewNHSAppTemplate/server-actions';
-import { FormState } from '../../utils/types';
-import { mainServerAction } from '../../app/create-template/main-server-action';
-import { getMockFormData } from '../helpers';
+import { FormState } from '../../../utils/types';
+import { mainServerAction } from '../../../app/create-template/main-server-action';
+import { getMockFormData } from '../../helpers';
 
 jest.mock('@forms/ReviewNHSAppTemplate/server-actions');
